chore: update base images for some components (#132)
This commit updates the base golang builder image for the cloudsqlproxy,
createdb, and createtree components. This uses the OpenShift brew
registry's Golang 1.21 builder. It is likely that all of the other
Dockerfiles in this repo and in others for the RHTAS project will need
to have their base images adjusted in the same way, and the github
action for image sha updates should also include brew in its updates.

Fixes: https://issues.redhat.com/browse/SECURESIGN-210